// src/app/_dto/masters/landed-cost-perc/TblDLandCost.dto.ts
export class TblDLandCost {
  DLmCost_SysID: number;
  DLmCost_DrASysID: number;
  DLmCost_DrACode: string;
  DLmCost_DrAName: string;
  DLmCurr_SysID: number;
  DLmCurr_Code: string;
  DLmCurr_Name: string;
  DLmCurr_Rate: number;
  DLmCost_Perce: number;
  DLmCost_CrASysID: number;
  DLmCost_CrACode: string;
  DLmCost_CrAName: string;
  DLmCost_NarraSysID: number;
  DLmCost_NarraCode: string;
  DLmCost_AccNarraName: string;


  // DcLmCost_SysID: number;
  // DcLmCost_SingleCo_SysID: number;
  // DcLmCost_SingleCo_Code: string;
  // DcLmCost_SingleCo_Name: string;

}